+++ src/core/GlyphRasterizer.cpp
@@ -1,0 +1,32 @@
+/////////////////////////////////////////////////////////////////////////////////////////////////
+//
+//  Tencent is pleased to support the open source community by making tgfx available.
+//
+//  Copyright (C) 2025 THL A29 Limited, a Tencent company. All rights reserved.
+//
+//  Licensed under the BSD 3-Clause License (the "License"); you may not use this file except
+//  in compliance with the License. You may obtain a copy of the License at
+//
+//      https://opensource.org/licenses/BSD-3-Clause
+//
+//  unless required by applicable law or agreed to in writing, software distributed under the
+//  license is distributed on an "as is" basis, without warranties or conditions of any kind,
+//  either express or implied. see the license for the specific language governing permissions
+//  and limitations under the license.
+//
+/////////////////////////////////////////////////////////////////////////////////////////////////
+
+#include "GlyphRasterizer.h"
+
+namespace tgfx {
+GlyphRasterizer::GlyphRasterizer(int width, int height,
+                                 std::shared_ptr<ScalerContext> scalerContext, GlyphID glyphID,
+                                 bool fauxBold, std::unique_ptr<Stroke> stroke)
+    : ImageCodec(width, height, Orientation::LeftTop), scalerContext(std::move(scalerContext)),
+      glyphID(glyphID), fauxBold(fauxBold), stroke(std::move(stroke)) {
+}
+
+bool GlyphRasterizer::readPixels(const ImageInfo& dstInfo, void* dstPixels) const {
+  return scalerContext->readPixels(glyphID, fauxBold, stroke.get(), dstInfo, dstPixels);
+}
+}  // namespace tgfx
